up msvc warning level and resolve warnings on test_tport
darcs-hash:20061122015202-16063-184d997dc7201f8ca6f3cc5f891c4f86df93e29e.gz

+++ libsofia-sip-ua/msg/test_class.h
@@ -53,7 +53,7 @@
 
 SOFIA_BEGIN_DECLS
 
-#define MSG_TEST_PROTOCOL_TAG ((void *)0xdeadbeef)
+#define MSG_TEST_PROTOCOL_TAG ((void *)(size_t)0xdeadbeef)
 
 #define MSG_TEST_VERSION_CURRENT "msg/1.0"
